Seperate truncate data-mig scripts for apps, folders and org
